requirements.txt
appmap>=1.1.0.dev0
pip~=21.0.1
wheel~=0.36.2
pytz~=2022.1
django~=3.2.14
asgiref~=3.4.1
sqlparse~=0.4.2
setuptools~=52.0.0
djangorestframework~=3.13.1